Renamed ant model manager.
